// src/models/types.ts
export interface ITableDataType {
	admittingProv: string
	bed: string
	bedStatus: string
	classF: string
	finClass: string
	room: string
	sex: string
	unitAbbr: string
	id: string,
	admitDate: string,
}

export interface IHeadCell {
	id: string,
	name: string,
}

export class abbrFilterType {
}

export class GraphDataType {
}

export class OrderType {
}

export interface TableDataType {
	id: string
	admittingProv: string
	bed: string
	bedStatus: string
	classF: string
	finClass: string
	room: string
	sex: string
	unitAbbr: string
	admitDate: string,
}

export class OrderByType {
}